requirements.txt
asgiref==3.2.7
autopep8==1.5.3
bcrypt==3.1.7
cffi==1.14.0
coverage==5.1
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
